refactor: Refatoração completa do status
